#49 - First step in implementing websockets - move regular HTTP stuff into its own namespace.
