+++ lib/smooth/application/network/http/regular/HTTPPacket.cpp
@@ -14,14 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-#include <smooth/application/network/http/HTTPPacket.h>
+#include <smooth/application/network/http/regular/HTTPPacket.h>
 
 #include <string>
 #include <unordered_map>
 #include <algorithm>
 #include <smooth/application/network/http/http_utils.h>
 
-namespace smooth::application::network::http
+namespace smooth::application::network::http::regular
 {
     HTTPPacket::HTTPPacket(ResponseCode code, const std::string& version,
                            const std::unordered_map<std::string, std::string>& new_headers,
