+++ lib/smooth/include/smooth/application/network/http/regular/IServerResponse.h
@@ -1,4 +1,4 @@
-// Smooth - C++ framework for writing applications based on Espressif's ESP-IDF.
+// Smooth - C++ framework for writing applications based on Espressif's ESP-IDF.smooth::application::network::http
 // Copyright (C) 2017 Per Malmberg (https://github.com/PerMalmberg)
 //
 // This program is free software: you can redistribute it and/or modify
@@ -16,14 +16,17 @@
 
 #pragma once
 
-#include <string>
+#include <memory>
+#include "responses/IRequestResponeOperation.h"
 
-namespace smooth::application::network::http
+namespace smooth::application::network::http::regular
 {
-    class ITemplateDataRetriever
+    class IServerResponse
     {
         public:
-            virtual ~ITemplateDataRetriever() = default;
-            virtual std::string get(const std::string& key) const = 0;
+            virtual ~IServerResponse() = default;
+
+            virtual void reply(std::unique_ptr<regular::responses::IRequestResponseOperation> response) = 0;
+            virtual void reply_error(std::unique_ptr<regular::responses::IRequestResponseOperation> response) = 0;
     };
 }
