+++ lib/smooth/include/smooth/application/network/http/regular/ITemplateDataRetriever.h
@@ -16,23 +16,14 @@
 
 #pragma once
 
-#include "responses/IRequestResponeOperation.h"
-#include "IServerResponse.h"
-#include <memory>
+#include <string>
 
-namespace smooth::application::network::http
+namespace smooth::application::network::http::regular
 {
-    class MIMEParser;
-
-    using RequestHandlerSignature = std::function<void(
-            IServerResponse& response,
-            const std::string& url,
-            bool first_part,
-            bool last_part,
-            const std::unordered_map<std::string, std::string>& headers,
-            const std::unordered_map<std::string, std::string>& request_parameters,
-            const std::vector<uint8_t>& content,
-            MIMEParser& mime
-    )>;
-
+    class ITemplateDataRetriever
+    {
+        public:
+            virtual ~ITemplateDataRetriever() = default;
+            virtual std::string get(const std::string& key) const = 0;
+    };
 }
