+++ lib/smooth/include/smooth/application/network/http/regular/MIMEParser.h
@@ -22,7 +22,7 @@
 #include <tuple>
 #include <unordered_map>
 
-namespace smooth::application::network::http
+namespace smooth::application::network::http::regular
 {
     class MIMEParser
     {
