+++ lib/smooth/include/smooth/application/network/http/regular/RequestHandlerSignature.h
@@ -16,14 +16,23 @@
 
 #pragma once
 
-namespace smooth::application::network::http
+#include "responses/IRequestResponeOperation.h"
+#include "IServerResponse.h"
+#include <memory>
+
+namespace smooth::application::network::http::regular
 {
-    enum class HTTPMethod
-    {
-            DELETE,
-            GET,
-            HEAD,
-            PUT,
-            POST
-    };
+    class MIMEParser;
+
+    using RequestHandlerSignature = std::function<void(
+            IServerResponse& response,
+            const std::string& url,
+            bool first_part,
+            bool last_part,
+            const std::unordered_map<std::string, std::string>& headers,
+            const std::unordered_map<std::string, std::string>& request_parameters,
+            const std::vector<uint8_t>& content,
+            MIMEParser& mime
+    )>;
+
 }
