+++ lib/smooth/include/smooth/application/network/http/regular/responses/FileContentResponse.h
@@ -16,17 +16,26 @@
 
 #pragma once
 
-#include <memory>
-#include "responses/IRequestResponeOperation.h"
+#include "StringResponse.h"
+#include <smooth/core/filesystem/Path.h>
+#include <smooth/core/filesystem/Fileinfo.h>
 
-namespace smooth::application::network::http
+namespace smooth::application::network::http::regular::responses
 {
-    class IServerResponse
+    class FileContentResponse
+            : public StringResponse
     {
         public:
-            virtual ~IServerResponse() = default;
+            explicit FileContentResponse(smooth::core::filesystem::Path full_path);
 
-            virtual void reply(std::unique_ptr<responses::IRequestResponseOperation> response) = 0;
-            virtual void reply_error(std::unique_ptr<responses::IRequestResponseOperation> response) = 0;
+            // Called at least once when sending a response and until ResponseStatus::AllSent is returned
+            ResponseStatus get_data(std::size_t max_amount, std::vector<uint8_t>& target) override;
+
+            void dump() const override;
+
+        private:
+            smooth::core::filesystem::Path path;
+            smooth::core::filesystem::FileInfo info;
+            std::size_t sent{0};
     };
 }
