+++ sdkconfig
@@ -194,6 +194,19 @@ CONFIG_ESP_TASK_WDT=y
 CONFIG_ESP_TASK_WDT_TIMEOUT_S=5
 CONFIG_ESP_TASK_WDT_CHECK_IDLE_TASK_CPU0=y
 CONFIG_ESP_TASK_WDT_CHECK_IDLE_TASK_CPU1=y
+CONFIG_ETH_USE_ESP32_EMAC=y
+CONFIG_ETH_PHY_INTERFACE_RMII=y
+# CONFIG_ETH_PHY_INTERFACE_MII is not set
+CONFIG_ETH_RMII_CLK_INPUT=y
+# CONFIG_ETH_RMII_CLK_OUTPUT is not set
+CONFIG_ETH_RMII_CLK_IN_GPIO=0
+CONFIG_ETH_SMI_MDC_GPIO=23
+CONFIG_ETH_SMI_MDIO_GPIO=18
+CONFIG_ETH_PHY_USE_RST=y
+CONFIG_ETH_PHY_RST_GPIO=5
+CONFIG_ETH_DMA_BUFFER_SIZE=512
+CONFIG_ETH_DMA_RX_BUFFER_NUM=10
+CONFIG_ETH_DMA_TX_BUFFER_NUM=10
 # CONFIG_ESP_EVENT_LOOP_PROFILING is not set
 CONFIG_ESP_EVENT_POST_FROM_ISR=y
 CONFIG_ESP_EVENT_POST_FROM_IRAM_ISR=y
@@ -222,12 +235,6 @@ CONFIG_ESP32_PHY_MAX_TX_POWER=20
 # CONFIG_ESP32_ENABLE_COREDUMP_TO_UART is not set
 CONFIG_ESP32_ENABLE_COREDUMP_TO_NONE=y
 # CONFIG_ESP32_ENABLE_COREDUMP is not set
-CONFIG_ETH_DMA_RX_BUF_NUM=10
-CONFIG_ETH_DMA_TX_BUF_NUM=10
-CONFIG_ETH_EMAC_L2_TO_L3_RX_BUF_MODE=y
-CONFIG_ETH_CHECK_LINK_STATUS_PERIOD_MS=2000
-CONFIG_ETH_EMAC_TASK_PRIORITY=20
-CONFIG_ETH_EMAC_TASK_STACK_SIZE=3072
 # CONFIG_FATFS_CODEPAGE_DYNAMIC is not set
 CONFIG_FATFS_CODEPAGE_437=y
 # CONFIG_FATFS_CODEPAGE_720 is not set
@@ -452,13 +459,14 @@ CONFIG_SPI_FLASH_DANGEROUS_WRITE_ABORTS=y
 CONFIG_SPI_FLASH_SUPPORT_ISSI_CHIP=y
 CONFIG_NETIF_IP_LOST_TIMER_INTERVAL=120
 CONFIG_TCPIP_LWIP=y
-CONFIG_SUPPRESS_SELECT_DEBUG_OUTPUT=y
-CONFIG_SUPPORT_TERMIOS=y
+CONFIG_VFS_SUPPRESS_SELECT_DEBUG_OUTPUT=y
+CONFIG_VFS_SUPPORT_TERMIOS=y
 CONFIG_SEMIHOSTFS_MAX_MOUNT_POINTS=1
 CONFIG_SEMIHOSTFS_HOST_PATH_MAX_LEN=128
 # CONFIG_WL_SECTOR_SIZE_512 is not set
 CONFIG_WL_SECTOR_SIZE_4096=y
 CONFIG_WL_SECTOR_SIZE=4096
+# CONFIG_WPA_ENTERPRISE is not set
 CONFIG_SMOOTH_MAX_MQTT_MESSAGE_SIZE=512
 CONFIG_SMOOTH_MAX_MQTT_OUTGOING_MESSAGES=10
 CONFIG_SMOOTH_MQTT_LOG_LEVEL_NONE=y
@@ -559,12 +567,6 @@ CONFIG_TASK_WDT_CHECK_IDLE_TASK_CPU1=y
 # CONFIG_EVENT_LOOP_PROFILING is not set
 CONFIG_POST_EVENTS_FROM_ISR=y
 CONFIG_POST_EVENTS_FROM_IRAM_ISR=y
-CONFIG_DMA_RX_BUF_NUM=10
-CONFIG_DMA_TX_BUF_NUM=10
-CONFIG_EMAC_L2_TO_L3_RX_BUF_MODE=y
-CONFIG_EMAC_CHECK_LINK_PERIOD_MS=2000
-CONFIG_EMAC_TASK_PRIORITY=20
-CONFIG_EMAC_TASK_STACK_SIZE=3072
 CONFIG_SUPPORT_STATIC_ALLOCATION=y
 # CONFIG_ENABLE_STATIC_TASK_CLEAN_UP_HOOK is not set
 CONFIG_TIMER_TASK_PRIORITY=1
@@ -599,4 +601,6 @@ CONFIG_SPI_FLASH_WRITING_DANGEROUS_REGIONS_ABORTS=y
 # CONFIG_SPI_FLASH_WRITING_DANGEROUS_REGIONS_FAILS is not set
 # CONFIG_SPI_FLASH_WRITING_DANGEROUS_REGIONS_ALLOWED is not set
 CONFIG_IP_LOST_TIMER_INTERVAL=120
+CONFIG_SUPPRESS_SELECT_DEBUG_OUTPUT=y
+CONFIG_SUPPORT_TERMIOS=y
 # End of deprecated options
