+++ test/http_server_test/SendBlob.h
@@ -16,12 +16,12 @@
 
 #pragma once
 
-#include <smooth/application/network/http/responses/StringResponse.h>
+#include <smooth/application/network/http/regular/responses/StringResponse.h>
 
 namespace http_server_test
 {
     class SendBlob
-            : public smooth::application::network::http::responses::StringResponse
+            : public smooth::application::network::http::regular::responses::StringResponse
     {
         public:
             explicit SendBlob(std::size_t amount);
@@ -29,7 +29,7 @@ namespace http_server_test
             ~SendBlob() override = default;
 
             // Called at least once when sending a response and until ResponseStatus::AllSent is returned
-            smooth::application::network::http::responses::ResponseStatus
+            smooth::application::network::http::regular::responses::ResponseStatus
             get_data(std::size_t max_amount, std::vector<uint8_t>& target) override;
 
         private:
