+++ test/http_server_test/http_server_test.h
@@ -24,7 +24,7 @@
 #include <smooth/core/network/SecureServerSocket.h>
 #include <smooth/application/network/http/HTTPServerClient.h>
 #include <smooth/application/network/http/HTTPProtocol.h>
-#include <smooth/application/network/http/ITemplateDataRetriever.h>
+#include <smooth/application/network/http/regular/ITemplateDataRetriever.h>
 
 #ifdef ESP_PLATFORM
 #include <smooth/core/filesystem/MMCSDCard.h>
